[Alpha] Adding support for validating venues (#63)
* Moving remarks on selected segment to a nicer spot

Moved the remarks to under "X segment(s) selected" instead of above it.

* Adding support for validating venues

Adding basic support for checking Venues like we currently check segments.
The only active check is a check if a venue has a city set.

Just blindingly copyied the functions of segments to venues. Will need some work of course, but for now this works.

* Removed duplicate code Venues

Removed the duplicate code surrounding the venues. Now we have a SimpleOBJECT that aims to match both Venue and Segment.

* Changes according to remarks by berestovskyy

Changes after the review.
Now I try to use a more dry approach

* Check if segment

Reversed the check