Fixes getting MPH vs KPH based on segment
Example PL: https://www.waze.com/nl/editor?env=usa&lon=-99.50151&lat=27.49976&zoom=5

This situation has the issue that a editor can edit in a country which is imperial (the US) and metrical (Mexico) what causes Validator to flag segments as having the wrong speed. This fixes that by looking at the country of a segment and converting the speed to mph or not according to a exception list.

The list of countries using MPH is a given and can be hardcoded imo.

This change was discused in #79