Releasing new alpha version (#84)
* Check brand in name in lowercase

Don't mind if the brand name has different case in the name.

* Removed unneeded condition

* Check for category OTHER

The OTHER category is a confussing "dump everything in this" category. Advice people against using it.

* Convert hardcoded check to param

This moves the hardcoded regexp to the params of the check.

* Added checks for phone numbers and URL

* Added checks for area vs point places

This information is taken from the Global Wazeopedia page (https://wazeopedia.waze.com/wiki/Global/Places#When_to_use_Area_or_Point) and should be localized per country when needed.

* Added new check for lock level on a venue

Some categories should be locked to a certain level. This checks for that.

* Added lock level 3  check

This information was taken from the Global wiki

* Monitor venue changes

Like segments, monitor venues for changes

* Correct regexp for URL checking

This one seems to work better

* Added Dutch trnaslations

Added Dutch translations for Venues and Speedlimits

* New Alpha Release

New release with new checks and changes