[Venues] Add support for venues checks (#89)
This closes #57 

* Adding support for validating venues

Adding basic support for checking Venues like we currently check segments.
The only active check is a check if a venue has a city set.

Just blindingly copyied the functions of segments to venues. Will need some work of course, but for now this works.

* [Alpha] Adding support for validating venues (#63)

* Moving remarks on selected segment to a nicer spot

Moved the remarks to under "X segment(s) selected" instead of above it.

* Adding support for validating venues

Adding basic support for checking Venues like we currently check segments.
The only active check is a check if a venue has a city set.

Just blindingly copyied the functions of segments to venues. Will need some work of course, but for now this works.

* Removed duplicate code Venues

Removed the duplicate code surrounding the venues. Now we have a SimpleOBJECT that aims to match both Venue and Segment.

* Changes according to remarks by berestovskyy

Changes after the review.
Now I try to use a more dry approach

* Check if segment

Reversed the check

* Added new venue checks

* Adding configuration option for venues

Master switch if a user wants to check venues or not.

* Adding categoryAttributes to the wme-externs

* Optimized checks

* Adjusted check after review

* Revert fix on check lotType

* Fix syntax error (will squash on merge)

* Minor fixes (#78)

* Add getPoint to externs
Plus other minor fixes

* Update i18n reference

* New alpha version + small fix (#80)

* New ALPHA version release

* Correct to report the venue, not the last segment

* Check brand in name in lowercase

Don't mind if the brand name has different case in the name.

* Removed unneeded condition

* Check for category OTHER

The OTHER category is a confussing "dump everything in this" category. Advice people against using it.

* Convert hardcoded check to param

This moves the hardcoded regexp to the params of the check.

* Added checks for phone numbers and URL

* Added checks for area vs point places

This information is taken from the Global Wazeopedia page (https://wazeopedia.waze.com/wiki/Global/Places#When_to_use_Area_or_Point) and should be localized per country when needed.

* Added new check for lock level on a venue

Some categories should be locked to a certain level. This checks for that.

* Added lock level 3  check

This information was taken from the Global wiki

* Monitor venue changes

Like segments, monitor venues for changes

* Correct regexp for URL checking

This one seems to work better

* Added Dutch trnaslations

Added Dutch translations for Venues and Speedlimits

* New Alpha Release

New release with new checks and changes

* Releasing new alpha version (#84)

* Check brand in name in lowercase

Don't mind if the brand name has different case in the name.

* Removed unneeded condition

* Check for category OTHER

The OTHER category is a confussing "dump everything in this" category. Advice people against using it.

* Convert hardcoded check to param

This moves the hardcoded regexp to the params of the check.

* Added checks for phone numbers and URL

* Added checks for area vs point places

This information is taken from the Global Wazeopedia page (https://wazeopedia.waze.com/wiki/Global/Places#When_to_use_Area_or_Point) and should be localized per country when needed.

* Added new check for lock level on a venue

Some categories should be locked to a certain level. This checks for that.

* Added lock level 3  check

This information was taken from the Global wiki

* Monitor venue changes

Like segments, monitor venues for changes

* Correct regexp for URL checking

This one seems to work better

* Added Dutch trnaslations

Added Dutch translations for Venues and Speedlimits

* New Alpha Release

New release with new checks and changes

* Zoom in more to get all venues

* Corrected $brand set to null sometimes

* Added new street check for venues

* Renamed files to fix building on Linux

Case sensitive OS requires correct file names

* Ignore natural features for this check

* New alpha release

* Converted excepted categories to check options

* Changed release notes

* mend

* Fixing merge

* Fixing merge 2

* More fixes

* Prepare merge of new translations

Updated with latest changes from master in dev branch

* Removed 260 from checks and translations

* Removed 260 from NL translation

* New release - Venue checking added

New release with venue checking added. See changelog for details.